requirements.txt
torch>=2.0
pytorch-lightning>=2.0
transformers>=4.40
datasets>=3.0
trl>=0.15
tokenizers>=0.20
accelerate>=1.0
safetensors>=0.4
wandb>=0.16
numpy>=1.24
matplotlib>=3.7
seaborn>=0.13
networkx>=3.0
tqdm>=4.60
pyyaml>=6.0
